# swagger.json
{"openapi":"3.0.0","paths":{"/users/{id}":{"get":{"operationId":"getUserById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]},"put":{"operationId":"updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users":{"post":{"operationId":"createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/auth/register":{"post":{"operationId":"register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/resend-verification-email":{"post":{"operationId":"sendVerificationEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMailDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/email-verify/{token}":{"get":{"operationId":"verifyToken","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/forgot-password":{"post":{"operationId":"forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMailDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/reset-password/{token}":{"post":{"operationId":"resetPassword","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/refresh-token":{"post":{"operationId":"refreshToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/change-password":{"put":{"operationId":"changePassword","parameters":[{"name":":id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"Template project","description":"Template project API description","version":"1.0","contact":{}},"tags":[{"name":"template","description":""},{"name":"example","description":""},{"name":"draft","description":""},{"name":"senior","description":""}],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string"},"phone":{"type":"string"},"role":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"confirmPassword":{"type":"string"}},"required":["username","phone","role","email","password","confirmPassword"]},"UpdateUserDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"SendMailDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"password":{"type":"string"},"confirmPassword":{"type":"string"}},"required":["password","confirmPassword"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"ChangePasswordDto":{"type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"type":"string"},"confirmPassword":{"type":"string"}},"required":["currentPassword","newPassword","confirmPassword"]}}}}